requirements.txt
# base ----------------------------------------
numpy
opencv-python
Pillow
tqdm
pywin32
python-pptx

# text-------------------------------------
markovgen